Send move Hub command on input change
