+++ src/web/context/jaia-actions.ts
@@ -18,6 +18,7 @@ export enum JaiaActions {
     TOGGLE_HYDROPHONE = "TOGGLE_HYDROPHONE",
     TOGGLE_CONSTANT_HEADING_SELECT = "TOGGLE_CONSTANT_HEADING_SELECT",
     TOGGLE_SELECT_HUB_LOCATION = "TOGGLE_SELECT_HUB_LOCATION",
+    MOVE_HUB = "MOVE_HUB",
     ADD_RALLY_POINT = "ADD_RALLY_POINT",
     DELETE_RALLY_POINT = "DELETE_RALLY_POINT",
     SEND_RALLY_MISSION = "SEND_RALLY_MISSION",
